moved subject processing from test into UserDataController; run using `node test`
